Add build script and config.php to configure production == true | false
Now, there are two config files, config.dev.php and config.prod.php.
In git, config.php is a symlink to config.dev.php.
In the build script, config.php is replaced with the contents of
config.prod.php.
This way, it is harder to make the mistake of accidentally leaving
production === true in the finished, exported plugin zip.