Refactor 'saveTask' action to use async/await
This makes a slight difference when you are using devtools, because
when actions are executed, you see 'start' and 'end' events in the
timeline.

When using then() and catch() inside the action, the
'start' and 'end' events fire within a couple ms of each other.

In comparison, when you use await and try{} catch{} in the Action,
the 'end' event will fire 500 ms later, like you would expect.