Fix 'assessment' chart colors in Stud.IP 5.1
I noticed that the colors of the assessment chart were wrong.
Successful attempts were being shown in orange instead of blue.

We define our own CSS rules to color the chart in lernmodule.less.
Dev Tools showed that our rules were being overridden by the defaults
defined in 'chartist.scss', which contains the default CSS for Chartist.
The rules were of equal specificity, so the last rule loaded won --
in this case, the default from Chartist.

I think that in the move to Stud.IP 5.1, we are loading the CSS files
 in a different order than in 4.6.  We are now importing Chartist using
 STUDIP.loadChunk() at the time that our chart is rendered.
 This means it gets loaded later than our own stylesheet, which we
 import using $this->addStylesheet().  Therefore, the default rules
 now take priority over the ones we wrote ourselves in case of a
 specificity tie.

I have added !important to the three rules which were affected.
This fixes it for the moment.