WIP - Drag handles to move start/end time of interactions on timeline.
I have looked into it a little bit and have come to the
conclusion that html5 drag-drop api is not necessarily the
correct api to use here.  It's apparently not possible to set
the cursor during drag-drop to any arbitrary one.  It will
always follow 'dropEffect', which only lets you pick between
a few presets like 'copy', 'move' and 'none'.  I think the
cursor should be resize-ew.

On StackOverflow, I received the tip that setPointerCapture
might be of use.
https://developer.mozilla.org/en-US/docs/Web/API/Element/setPointerCapture

I think I will use mousedown/mouseup and setPointerCapture
to handle this feature.