Begin to re-implement file and folder picker components in Vue 3.
I have read a lot of code and thought a lot about how to reuse
the CoursewareFileChooser and CoursewareFolderChooser components
from the core of Stud.IP.  I have not found any easy solution.
So, I have chosen to look into re-implementing these components
in Vue 3 / Typescript, as I have previously done with the
TabComponent/TabsComponent from the Stud.IP core.

Here is a brain dump summarizing my thoughts from the past week
about the various factors hindering me from directly re-using these
components from the core.

There are multiple sources of incompatibility and technical debt:

- Vue 3 vs. Vue 2
Our Lernmodule editor and viewer apps are written in Vue 3.
The core is written in Vue 2.  There is no simple way to use
a Vue 3 component in a Vue 2 app.  We chose to use an iframe.

- Iframe
The iframe approach was easy to implement and provided a good
dev experience.  The iframe itself reloaded when the HMR demanded
it, while the surrounding courseware page did not have to reload.
But the iframe somewhat complicates the passing of data
between Vue 3 and Vue 2. The VueX store object from the core,
which is depended upon by the core file/folder picker components,
cannot be passed through an iframe boundary, so it cannot easily
be passed through to the plugin's Vue 3 app in order to make it
accessible to a folder/file picker within Vue 3.

- VueX 3 vs. Vue 3 incompatibility
VueX 3 is not advertised to be compatible with Vue 3.
It may be possible to use the two together to a certain extent,
maybe by writing our own compatibility shims and/or avoiding
certain constructs that are not compatible with Vue 3.
But it's possible that we would run into further compatibility
bugs later down the road.

- Singleton assumption in Lernmodule editor and viewer
Originally, the Lernmodule Vue.js components were intended to
be used only in the Lernmodule plugin of Stud.IP.
In the LernmodulePlugin, there is only one task edited at a time
on a given page, so we thought it safe to keep the state of the
editor in a central VueX store.
This worked well with the iframe approach that we chose to use,
as outlined above, to integrate the LM editor/viewer into the
Courseware.
But this presents a hinderance if we intend to refactor our plugin
to mount our Vue 3 components directly in the DOM tree of the core
Vue 2 app, allowing direct access to the core VueX store.
We would need to refactor to remove the assumption that there is
 only a single LM editor instance in the app, as multiple editors
 (multiple instances of a given LM CW block type) would need to
  be able to co-exist in one app.