Refactor: Create component 'ImageWithHotspots'
This component serves a role analogous to VideoPlayer from
Interactive Video.
It can be used in both the editor and the viewer.
This saves us the trouble of duplicating code between
the viewer/editor. E.g. to position the image and position
the hotspots.
(It's critical that the hotspots are positioned the same way
between the viewer/editor. We don't want to implement that
twice in two subtly different ways. That way lie bugs :)